refactor: replace crate `tokio` to `async-fs` and `async-io` (#164)
BREAKING CHANGE: `download::download_files` is renamed to
`download::download_concurrent_and_check` and
`download::download_concurrent`

+++ crates/config/src/download.rs
@@ -53,7 +53,7 @@ pub struct DownloadConfig {
     ///
     /// A value of `0` disables throttling (unlimited speed).
     #[serde(default)]
-    pub max_download_speed: usize,
+    pub max_download_speed: u64,
 
     /// Custom mirror configuration.
     ///
